+++ src/storage/mod.rs
@@ -7,7 +7,7 @@ mod stored_value;
 mod value;
 
 pub(crate) use in_memory::{
-    ExpirationUpdate, ExpireCondition, InMemoryStore, SetCondition, SetExpiration,
+    ExpirationUpdate, ExpireCondition, InMemoryStore, SetCondition, SetExpiration, SetOperation,
 };
 pub(crate) use snapshot::{SnapshotDataError, SnapshotEntry, SnapshotValue};